Changed my mind, allowing optional default properties
I don't think this is that useful, but for consistency I thought I should attempt to normalize
properties as well.  I had hoped to figure out a way that it could detect if properties should
be there but weren't.  But I couldn't figure out an intuitive way to do that.